Replace psutil with pgrep, virtualenv with venv
Psutil is used in master.py to check if a process with a given name is
running. Installing psutil is messy and using it is messier still. This
commit replaces it with pgrep, a cleaner alternative.

Virtualenv is used prior to this commit to initialize a python virtual
environment on the nodes. This was replaced with venv, which worked more
easily for me.