Add helper to communicate with specific node
This commit add the MqttSocket class. Much like a websocket allows open
two way communication, an MqttSocket allows a user to define a "listen"
and a "send" topic, which should be different. They can then do
a blocking operation which sends a request to the target topic and
awaits a response on the listen topic.

A good use of this helper is requesting data from a specific node, which
follows a "request / response" pattern.