Add functions for webapp to post updates
Located in help_lib.py, there are 4 new functions:
- addInteraction
- deleteInteraction
- updateInteraction
- updateNode

On the surface, these are what the webapp should call to update the
state of the system.

Internally, it works by publishing to the topic 'webapp/updates'.
Currently the nodes have not been set up to parse this, but once they
are, all nodes connected to the broker should update their internal
state when they receive such a message.