Define heartbeat behavior
This commit adds an on_disconnect function, invoked if the broker
unexpectedly disconnects from the node. This implies the master node has
gone down.

It also adds a will for each node, published to the 'heartbeats' topic.
If a node unexpectedly disconnects from the broker, it will send the
message to all other nodes, allowing them to update their internal state
that the given other node has gone down.